Move .h and .cpp files to Correct Location
Visual Studio cannot read the files from outside the folder
