--- Door.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Entity.h"
-#include <iostream>
-
-class Door : public Entity
-{
-private:
-
-public:
-	Door(int y_pos, int x_pos);
-};
