--- Key.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Entity.h"
-#include <iostream>
-
-class Key : public Entity
-{
-private:
-
-public:
-	Key(int y_pos, int x_pos);
-};
