--- Player.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "Entity.h"
-#include <iostream>
-
-class Player : public Entity
-{
-private:
-
-public:
-	Player(int y_pos, int x_pos);
-};
