+++ SP1Framework/Framework/Door.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "Entity.h"
+#include <iostream>
+
+class Door : public Entity
+{
+private:
+
+public:
+	Door(int y_pos, int x_pos);
+};
