+++ SP1Framework/Key.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "Entity.h"
+#include <iostream>
+
+class Key : public Entity
+{
+private:
+
+public:
+	Key(int y_pos, int x_pos);
+};
