+++ SP1Framework/Player.h
@@ -1,0 +1,11 @@
+#pragma once
+#include "Entity.h"
+#include <iostream>
+
+class Player : public Entity
+{
+private:
+
+public:
+	Player(int y_pos, int x_pos);
+};
